Nest stratiform's Tel handle and openables, but not TelFlag
`TelHandle`, `TelOpenable` and `TelViewOpenable` become `Tel.Handle`,
`Tel.Openable` and `Tel.ViewOpenable`. The two openables could not be eliminated
into their givens as R7 suggests — a comment on the donor records that an
anonymous subclass freshens the handle's capability type in the inferred `Result`
member — so nesting is the only move available.

`Tel.Openable` shadows the `aperture.Openable` it extends, which is why
`ZipOpenable` was reverted in the first pass. It does not have to be: qualifying
the base as `extends aperture.Openable` fixes it, exactly as a nested `Error`
qualifies `fulminate.Error`. The Zip pair is worth revisiting on the same basis.

`TelFlag` stays whole. `Tel.Flag` is ambiguous with `stratiform.Tels.Flag`, a
distinct schema-level type, at every `import Tels.*` site inside `object Tel` —
a collision with a *sibling's* member that checking the outer object for the
target name does not reveal.
